requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
azure-identity==1.15.0
azure-ai-projects==1.0.0b4
requests==2.31.0
gunicorn==21.2.0
